# Conflicts:
#	.devops/code-review-pipelines.yml

+++ src/test/java/it/gov/pagopa/common/mongo/singleinstance/SingleEmbeddedMongodbAutoConfiguration.java
@@ -1,17 +1,23 @@
 package it.gov.pagopa.common.mongo.singleinstance;
 
+import de.flapdoodle.embed.mongo.commands.MongoImportArguments;
 import de.flapdoodle.embed.mongo.commands.MongodArguments;
 import de.flapdoodle.embed.mongo.config.Net;
 import de.flapdoodle.embed.mongo.distribution.IFeatureAwareVersion;
-import de.flapdoodle.embed.mongo.spring.autoconfigure.*;
+import de.flapdoodle.embed.mongo.spring.autoconfigure.AbstractServerFactory;
+import de.flapdoodle.embed.mongo.spring.autoconfigure.EmbeddedMongoAutoConfiguration;
+import de.flapdoodle.embed.mongo.spring.autoconfigure.MongodWrapper;
+import de.flapdoodle.embed.mongo.spring.autoconfigure.ReactiveClientServerFactory;
+import de.flapdoodle.embed.mongo.spring.autoconfigure.SyncClientServerFactory;
 import de.flapdoodle.embed.mongo.transitions.Mongod;
+import com.mongodb.MongoClientSettings;
 import it.gov.pagopa.common.mongo.EmbeddedMongodbTestClient;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration;
-import org.springframework.boot.autoconfigure.mongo.MongoProperties;
-import org.springframework.boot.autoconfigure.mongo.MongoReactiveAutoConfiguration;
+import org.springframework.boot.mongodb.autoconfigure.MongoAutoConfiguration;
+import org.springframework.boot.mongodb.autoconfigure.MongoProperties;
+import org.springframework.boot.mongodb.autoconfigure.MongoReactiveAutoConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.env.ConfigurableEnvironment;
@@ -22,6 +28,7 @@ import org.springframework.test.context.event.annotation.AfterTestClass;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
@@ -47,7 +54,7 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
         if(SingleInstanceMongodWrapper.singleMongodNet!=null){
             ConfigurableEnvironment env = context.getEnvironment();
             env.getPropertySources().addFirst(new MapPropertySource("embeddedMongoReusedProperties",
-                    Map.of("spring.data.mongodb.port", SingleInstanceMongodWrapper.singleMongodNet.getPort())));
+                    Map.of("spring.mongodb.port", SingleInstanceMongodWrapper.singleMongodNet.getPort())));
             super.net(context);
 
             return SingleInstanceMongodWrapper.singleMongodNet;
@@ -58,7 +65,9 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
 
     @AfterTestClass
     void clearData(){
-        embeddedMongodbTestClient.dropDatabase();
+        if (SingleInstanceMongodWrapper.singleMongodNet != null) {
+            embeddedMongodbTestClient.dropDatabase();
+        }
     }
 
     @ConditionalOnClass(name = {
@@ -71,7 +80,7 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
 
         ReactiveClientServerWrapperConfig() {
             try {
-                unprotectedReactiveClientServerFactoryConstructor = ReactiveClientServerFactory.class.getDeclaredConstructor(MongoProperties.class);
+                unprotectedReactiveClientServerFactoryConstructor = ReactiveClientServerFactory.class.getDeclaredConstructor(MongoProperties.class, MongoClientSettings.class);
                 unprotectedReactiveClientServerFactoryConstructor.setAccessible(true);
             } catch (NoSuchMethodException e) {
                 throw new IllegalStateException("Cannot unprotect AbstractServerFactory constructor", e);
@@ -83,10 +92,10 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
                 destroyMethod = "stop"
         )
         @ConditionalOnMissingBean
-        public MongodWrapper reactiveClientServerWrapper(IFeatureAwareVersion version, MongoProperties properties, Mongod mongod, MongodArguments mongodArguments) {
+        public MongodWrapper reactiveClientServerWrapper(IFeatureAwareVersion version, MongoProperties properties, MongoClientSettings mongoClientSettings, Mongod mongod, MongodArguments mongodArguments, List<MongoImportArguments> mongoImportArguments) {
             return Objects.requireNonNullElseGet(
                     singleMongodWrapperInstance,
-                    () -> createMongodWrapper(unprotectedReactiveClientServerFactoryConstructor, version, properties, mongod, mongodArguments));
+                    () -> createMongodWrapper(unprotectedReactiveClientServerFactoryConstructor, version, properties, mongoClientSettings, mongod, mongodArguments, mongoImportArguments));
         }
 
         @Bean
@@ -106,7 +115,7 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
 
         SyncClientServerWrapperConfig() {
             try {
-                unprotectedSyncClientServerFactoryConstructor = SyncClientServerFactory.class.getDeclaredConstructor(MongoProperties.class);
+                unprotectedSyncClientServerFactoryConstructor = SyncClientServerFactory.class.getDeclaredConstructor(MongoProperties.class, MongoClientSettings.class);
                 unprotectedSyncClientServerFactoryConstructor.setAccessible(true);
             } catch (NoSuchMethodException e) {
                 throw new IllegalStateException("Cannot unprotect AbstractServerFactory constructor", e);
@@ -118,10 +127,10 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
                 destroyMethod = "stop"
         )
         @ConditionalOnMissingBean
-        public MongodWrapper syncClientServerWrapper(IFeatureAwareVersion version, MongoProperties properties, Mongod mongod, MongodArguments mongodArguments) {
+        public MongodWrapper syncClientServerWrapper(IFeatureAwareVersion version, MongoProperties properties, MongoClientSettings mongoClientSettings, Mongod mongod, MongodArguments mongodArguments, List<MongoImportArguments> mongoImportArguments) {
             return Objects.requireNonNullElseGet(
                     singleMongodWrapperInstance,
-                    () -> createMongodWrapper(unprotectedSyncClientServerFactoryConstructor, version, properties, mongod, mongodArguments));
+                    () -> createMongodWrapper(unprotectedSyncClientServerFactoryConstructor, version, properties, mongoClientSettings, mongod, mongodArguments, mongoImportArguments));
         }
 
         @Bean
@@ -131,9 +140,9 @@ public class SingleEmbeddedMongodbAutoConfiguration extends EmbeddedMongoAutoCon
         }
     }
 
-    private static MongodWrapper createMongodWrapper(Constructor<? extends AbstractServerFactory<?>> unprotectedSyncClientServerFactoryConstructor, IFeatureAwareVersion version, MongoProperties properties, Mongod mongod, MongodArguments mongodArguments) {
+    private static MongodWrapper createMongodWrapper(Constructor<? extends AbstractServerFactory<?>> unprotectedSyncClientServerFactoryConstructor, IFeatureAwareVersion version, MongoProperties properties, MongoClientSettings mongoClientSettings, Mongod mongod, MongodArguments mongodArguments, List<MongoImportArguments> mongoImportArguments) {
         try {
-            return singleMongodWrapperInstance = new SingleInstanceMongodWrapper(unprotectedSyncClientServerFactoryConstructor.newInstance(properties).createWrapper(version, mongod, mongodArguments));
+            return singleMongodWrapperInstance = new SingleInstanceMongodWrapper(unprotectedSyncClientServerFactoryConstructor.newInstance(properties, mongoClientSettings).createWrapper(version, mongod, mongodArguments, mongoImportArguments));
         } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
             throw new IllegalStateException("Cannot call protected constructor", e);
         }
